small changes XD

+++ src/main/java/julianf/dev/anamar/product/ProductRepository.java
@@ -2,5 +2,5 @@ package julianf.dev.anamar.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Short> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
